#55 searchbarを分割、現状動作していない。
