+++ iOSEngineerCodeCheck/Views/Base.lproj/SearchPage.storyboard
@@ -8,15 +8,15 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--GitHubリポジトリ検索App-->
+        <!--TableViewAndSearchBar-->
         <scene sceneID="0Yw-Vc-k2Q">
             <objects>
-                <tableViewController storyboardIdentifier="SearchViewController" useStoryboardIdentifierAsRestorationIdentifier="YES" id="MOh-CZ-3ki" customClass="SearchViewController" customModule="iOSEngineerCodeCheck" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController storyboardIdentifier="SearchViewController" useStoryboardIdentifierAsRestorationIdentifier="YES" id="MOh-CZ-3ki" userLabel="TableViewAndSearchBar" customClass="SearchViewController" customModule="iOSEngineerCodeCheck" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" sectionFooterHeight="28" id="Fpt-Ev-QNW" customClass="TableViewClass" customModule="iOSEngineerCodeCheck" customModuleProvider="target">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                        <searchBar key="tableHeaderView" contentMode="redraw" id="6rq-CD-Hob">
+                        <searchBar key="tableFooterView" contentMode="redraw" text="" id="6rq-CD-Hob" customClass="UISearchBarClass" customModule="iOSEngineerCodeCheck" customModuleProvider="target">
                             <rect key="frame" x="0.0" y="0.0" width="414" height="44"/>
                             <autoresizingMask key="autoresizingMask" widthSizable="YES" flexibleMaxY="YES"/>
                             <textInputTraits key="textInputTraits"/>
@@ -28,7 +28,7 @@
                     </tableView>
                     <navigationItem key="navigationItem" title="GitHubリポジトリ検索App" id="ktq-eC-ZBq"/>
                     <connections>
-                        <outlet property="searchBar" destination="6rq-CD-Hob" id="3gq-mK-4M3"/>
+                        <outlet property="searchBar" destination="Fpt-Ev-QNW" id="9ye-MF-KfV"/>
                         <outlet property="tableView" destination="Fpt-Ev-QNW" id="4dE-oy-abM"/>
                     </connections>
                 </tableViewController>
